Update LoraHelperViewModel.cs
The problem is in the LoadAsync() method. Looking at the flow:
1.	IsLoading = true is set at the beginning
2.	Various async operations happen
3.	await LoadNextPageAsync() is called
4.	StartIndexing() is called (which starts a background task)
5.	await Dispatcher.UIThread.InvokeAsync(() => IsLoading = false) is called
The issue is that IsLoading is being set to false too early - right after StartIndexing() and before all the card loading is actually complete. The LoadNextPageAsync() method loads cards in pages, and if there are many cards, this might still be running when IsLoading is set to false.